refactor: reorganize screens structure
