+++ src/screens/MotoDetailsScreen/index.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
-import { View, Text, StyleSheet, Image } from 'react-native';
+import { View, Text, Image } from 'react-native';
 import { StackScreenProps } from '@react-navigation/stack';
+import { styles } from './styles';
 
 type RootStackParamList = {
   MotoDetails: {
@@ -41,35 +42,3 @@ export default function MotoDetailsScreen({ route }: Props) {
   );
 }
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    padding: 24,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-  },
-  title: {
-    fontSize: 24,
-    fontWeight: 'bold',
-    marginBottom: 16,
-  },
-  motoImage: {
-    width: 200,
-    height: 120,
-    resizeMode: 'contain',
-    marginBottom: 24,
-  },
-  detailsContainer: {
-    width: '100%',
-  },
-  detailLabel: {
-    fontSize: 16,
-    fontWeight: '600',
-    color: '#169BA4',
-    marginTop: 12,
-  },
-  detailValue: {
-    fontSize: 18,
-    fontWeight: '400',
-  },
-});
