+++ src/screens/SettingsScreen/index.tsx
@@ -1,7 +1,8 @@
 import React from 'react';
-import { View, Text, StyleSheet, Switch, TouchableOpacity, Image, Alert } from 'react-native';
+import { View, Text, Switch, TouchableOpacity, Image, Alert } from 'react-native';
 import AsyncStorage from '@react-native-async-storage/async-storage';
 import { useNavigation } from '@react-navigation/native';
+import { styles } from './styles';
 
 export default function SettingsScreen() {
   const [notificationsEnabled, setNotificationsEnabled] = React.useState(true);
@@ -38,45 +39,3 @@ export default function SettingsScreen() {
   );
 }
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    padding: 24,
-    backgroundColor: '#fff',
-  },
-  profileContainer: {
-    alignItems: 'center',
-    marginBottom: 24,
-  },
-  profileImage: {
-    width: 72,
-    height: 72,
-    borderRadius: 36,
-    marginBottom: 8,
-  },
-  profileName: {
-    fontSize: 18,
-    fontWeight: '600',
-  },
-  settingItem: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    marginBottom: 16,
-  },
-  settingText: {
-    fontSize: 18,
-  },
-  logoutButton: {
-    marginTop: 40,
-    backgroundColor: '#169BA4',
-    paddingVertical: 14,
-    borderRadius: 6,
-    alignItems: 'center',
-  },
-  logoutButtonText: {
-    color: '#fff',
-    fontSize: 16,
-    fontWeight: '600',
-  },
-});
