+++ src/screens/WelcomeScreen/index.tsx
@@ -2,15 +2,15 @@ import React from "react";
 import {
   View,
   Text,
-  StyleSheet,
   Image,
   TouchableOpacity,
   SafeAreaView,
 } from "react-native";
 import { StackNavigationProp } from "@react-navigation/stack";
 import { useNavigation } from "@react-navigation/native";
+import { styles } from "./styles";
 
-const solutionImage = require("../../assets/SolutionsNote.png");
+const solutionImage = require("../../../assets/SolutionsNote.png");
 
 type RootStackParamList = {
   Welcome: undefined;
@@ -49,46 +49,3 @@ export default function WelcomeScreen() {
   );
 }
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: "#fff",
-    justifyContent: "flex-end",
-    paddingHorizontal: 20,
-    paddingBottom: 40,
-  },
-  logoContainer: {
-    flex: 1,
-    justifyContent: "center",
-    alignItems: "center",
-  },
-  logo: {
-    width: 180,
-    height: 180,
-    resizeMode: "contain",
-  },
-  primaryButton: {
-    backgroundColor: "#169BA4",
-    borderRadius: 6,
-    paddingVertical: 15,
-    alignItems: "center",
-    marginBottom: 10,
-  },
-  primaryButtonText: {
-    color: "#fff",
-    fontSize: 18,
-    fontWeight: "600",
-  },
-  secondaryButton: {
-    borderColor: "#169BA4",
-    borderWidth: 1,
-    borderRadius: 6,
-    paddingVertical: 15,
-    alignItems: "center",
-  },
-  secondaryButtonText: {
-    color: "#169BA4",
-    fontSize: 18,
-    fontWeight: "600",
-  },
-});
