+++ src/screens/WelcomeScreen/styles.ts
@@ -1,0 +1,45 @@
+import { StyleSheet } from 'react-native';
+
+export const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#fff',
+    justifyContent: 'flex-end',
+    paddingHorizontal: 20,
+    paddingBottom: 40,
+  },
+  logoContainer: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+  logo: {
+    width: 180,
+    height: 180,
+    resizeMode: 'contain',
+  },
+  primaryButton: {
+    backgroundColor: '#169BA4',
+    borderRadius: 6,
+    paddingVertical: 15,
+    alignItems: 'center',
+    marginBottom: 10,
+  },
+  primaryButtonText: {
+    color: '#fff',
+    fontSize: 18,
+    fontWeight: '600',
+  },
+  secondaryButton: {
+    borderColor: '#169BA4',
+    borderWidth: 1,
+    borderRadius: 6,
+    paddingVertical: 15,
+    alignItems: 'center',
+  },
+  secondaryButtonText: {
+    color: '#169BA4',
+    fontSize: 18,
+    fontWeight: '600',
+  },
+});
